gemini-meet-notes-dev/pyproject.toml
[project]
name = "gemini-meet-notes-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "google-cloud-firestore",
    "python-dotenv",
    "pydantic"
]
